+++ CHANGELOG.md
@@ -26,7 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `IssNavisworksBuilder` with `EnableUserAppData` to install in current `UserAppData`.
 - Update `IssAppBundleBuilder` to use `PrivilegesRequired.Lowest` when install with `UserAppData`.
 ### Tests
-- Add `AutoCADExtension` tests with `AutoCADAddin` files for version `2019`, `2021` and `2025`.
+- Add `AutoCADExtension` tests with `AutoCADAddin` files for version `2019` to `2026`.
 - Add `NavisworksExtension` tests with `NavisworksAddin` files for version `2021` to `2026`.
 
 ## [1.8.3] / 2025-03-20